Rudimentary tracking/searching for runtime parameter changes
Working on rudimentary tracking and search for runtime (acquisition) parameters.  First step is working with parameters that are already available in the coverage plotter for colormap use.

'Parameters' and 'Search' tabs have been added to start laying out a text log of initial settings and all subsequent changes in the loaded/plotted data.  Basic search is supported for identifying the times of changes to parameters (user selected, printed to screen).  Next steps will be to fine-tune the search for identifying times associated with specific settings (e.g., when system was in Very Shallow mode, not just changes to 'Depth Mode' in general).